+++ packages/loopover-engine/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@loopover/engine",
-  "version": "3.15.3",
+  "version": "3.16.0",
   "license": "AGPL-3.0-only",
   "type": "module",
   "description": "Shared deterministic engine logic for the LoopOver review stack and loopover-miner.",
