+++ packages/loopover-mcp/package.json
@@ -46,7 +46,7 @@
   },
   "dependencies": {
     "@loopover/contract": "^0.1.0",
-    "@loopover/engine": "^3.15.2",
+    "@loopover/engine": "^3.16.0",
     "@modelcontextprotocol/sdk": "1.29.0",
     "posthog-node": "^5.46.1",
     "zod": "^4.4.3"
