+++ packages/loopover-miner/expected-engine.version
@@ -1,1 +1,1 @@
-3.15.3
+3.16.0
